Port to TypeScript
