+++ packages/version-from-git/src/tsconfig.json
@@ -2,7 +2,6 @@
 {
   "compilerOptions": {
     "esModuleInterop": true,
-    "jsx": "react",
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
     "noEmit": true,
