+++ packages/version-from-git/src/tsconfig.precommit.production.json
@@ -1,7 +1,6 @@
 {
   "compilerOptions": {
     "esModuleInterop": true,
-    "jsx": "react",
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
     "noEmit": true,
