+++ packages/version-from-git/tsup.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from 'tsup';
+
+export default defineConfig([
+  {
+    entry: {
+      'version-from-git': './src/index.mts'
+    },
+    format: ['esm'],
+    sourcemap: true
+  }
+]);
